Replaced all mentioned of "signalfx" with "splunk"
Except README, which is taken care of separately.
