+++ custom/src/main/resources/META-INF/services/io.opentelemetry.javaagent.bootstrap.spi.TracerCustomizer
@@ -1,1 +1,1 @@
-com.signalfx.opentelemetry.InstrumentationLibraryTracerCustomizer
+com.splunk.opentelemetry.InstrumentationLibraryTracerCustomizer
