+++ smoke-tests/src/test/resources/logback.xml
@@ -12,6 +12,6 @@
   </root>
 
   <logger name="io.opentelemetry" level="debug"/>
-  <logger name="com.signalfx" level="debug"/>
+  <logger name="com.splunk" level="debug"/>
 
 </configuration>
